q3-earnings-s3-upload/requirements.txt
requests
pydantic
psycopg2-binary
boto3
python-dotenv
apache-airflow
